Catch error if user sitepermissions don't exist.
For the most part, if not using `SitePermissionMiddleware` Mezzanine
falls back on the `is_staff` attribute seamlessly. But since
d5d21ba527bd4 it's possible to have users without sitepermissions. This
breaks the admin dropdown, but not much else. This fix should allow
single site projects to continue to leave `SitePermissionMiddleware`
uninstalled.