Fixed login form to not use request.REQUEST
This is not available in Django 1.9 and greater, so without this fix the
forms are (slightly) broken.  There doesn't seem to be any reason to use
REQUEST instead of POST - the form is a POST one, and the parameters are not
used elsewhere in the code base to construct a querystring.